启动爬虫任务功能完成

+++ server/model/dcncy/spider/spider_task.go
@@ -5,7 +5,7 @@ import (
 )
 
 const (
-	Unprocessed int8 = iota
+	Unprocessed int = iota
 	Processing
 	Completed
 	Failed
